# Pre-compile yaml file into py files