--- a/tox.ini
+++ b/tox.ini
@@ -5,4 +5,5 @@ envlist = py26, py27, pypy, py31, py32, py33, py34, py35, docs, pep8, py3pep8
 deps =
     pyyaml
 commands =
+    python setup.py develop
     python ua_parser/user_agent_parser_test.py