--- a/ua_parser/user_agent_parser.py
+++ b/ua_parser/user_agent_parser.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python
-#
 # Copyright 2009 Google Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the 'License')