This is a tremendous speedup vs parsing JSON

New code:

```
$ python -m timeit -s 'from ua_parser import user_agent_parser' 'reload(user_agent_parser)'
10000 loops, best of 3: 346 usec per loop
```

Old:

```
$ python -m timeit -s 'from ua_parser import user_agent_parser' 'reload(user_agent_parser)'
10 loops, best of 3: 151 msec per loop
```

This is multiple orders of magnitude faster, and even faster on slower computers.

151000 usec vs 346 usec. :)

Fixes #23 #35